requirements.txt
python-dotenv==1.0.0
openai==1.3.0
gTTS==2.3.2
sounddevice==0.4.6
soundfile==0.12.1
python-decouple==3.8
